Added image to home page
