+++ src/App.jsx
@@ -5,12 +5,10 @@ import GrommetApp from 'grommet/components/App';
 import Box from 'grommet/components/Box';
 
 import Nav from 'components/Nav.jsx';
-import Home from 'components/Home.jsx';
-import Partners from 'components/Partners.jsx';
-import Joining from 'components/Joining.jsx';
-import Forms from 'components/Forms.jsx';
+import Home from 'components/home/Home.jsx';
+import About from 'components/about/About.jsx';
+import GetInvolved from 'components/get-involved/GetInvolved.jsx';
 import Parents from 'components/Parents.jsx';
-import Team from 'components/Team.jsx';
 
 const App = () => (
   <HashRouter>
@@ -32,11 +30,9 @@ const App = () => (
 
         {/* Set the children of the primary component */}
         <Route path='/home' component={Home} />
-        <Route path='/partners' component={Partners} />
-        <Route path='/join' component={Joining} />
-        <Route path='/forms' component={Forms} />
+        <Route path='/about' component={About} />
+        <Route path='/get-involved' component={GetInvolved} />
         <Route path='/parents' component={Parents} />
-        <Route path='/team' component={Team} />
       </Box>
     </GrommetApp>
   </HashRouter>
